Guard against invalid github remote
If "github" is specified in a dist.ini, but a user doesn't
have a github remote defined, the url was wrongly coming
back as just "github".  This patch skips repository detection
when the requested git remote appears to be invalid and issues
a warnings.

This seems the most "user friendly" for when contributors have
a different repository setup than the primary author